BaseTabbedDialog has Leftnavigation as option
- Refactoring: instead of putting the left navigation style into efaConfigDialog only, it is put into BaseTabbedDialog, so that any other dialog may use this navigation instead of tabbedpane navigation.
- left navigation is an option in baseTabbedDialog now
- can be activated by setnavigationMode(MODE_LEFT_NAVIGATION)
- Statistics Dialog uses this pattern now
